requirements.txt
selene~=2.0.0rc9
pytest~=8.3.3
allure-pytest


selenium~=4.27.1
python-dotenv~=1.0.1
requests~=2.32.3
jsonschema~=4.23.0